common base class for ModConfigMenu and OwmlConfigMenu. removes hacky code
