+++ OWML.Common/Menus/IModConfigMenu.cs
@@ -1,11 +1,7 @@
 ﻿namespace OWML.Common.Menus
 {
-    public interface IModConfigMenu : IModPopupMenu
+    public interface IModConfigMenu : IBaseConfigMenu
     {
-        IModManifest Manifest { get; }
         IModBehaviour Mod { get; }
-
-        void Initialize(Menu modMenuCopy, IModToggleInput toggleTemplate, IModSliderInput sliderTemplate, 
-            IModTextInput textInputTemplate, IModNumberInput numberInputTemplate, IModComboInput comboInputTemplate);
     }
 }
