+++ OWML.Common/Menus/IModMenus.cs
@@ -5,7 +5,7 @@
         IModMainMenu MainMenu { get; }
         IModPauseMenu PauseMenu { get; }
         IModsMenu ModsMenu { get; }
-        IModConfigMenu OwmlMenu { get; }
+        IBaseConfigMenu OwmlMenu { get; }
         IModInputMenu InputMenu { get; }
     }
 }
