+++ OWML.ModHelper.Menus/ModsMenu.cs
@@ -108,7 +108,7 @@ namespace OWML.ModHelper.Menus
             return modsTab;
         }
 
-        private void InitConfigMenu(IModConfigMenu modConfigMenu, IModTabbedMenu options)
+        private void InitConfigMenu(IBaseConfigMenu modConfigMenu, IModTabbedMenu options)
         {
             var toggleTemplate = options.InputTab.ToggleInputs[0];
             var sliderTemplate = options.InputTab.SliderInputs[0];
